Fixes for tests:
