+++ package.json
@@ -19,12 +19,12 @@
     "@lit-labs/ssr/lib/render-lit-html.js"
   ],
   "scripts": {
-    "build": "npm run dev:prepare && npm run prepack",
+    "build": "npm run dev:prepare && npm run prepack && npm run dev:build",
     "prepack": "nuxt-module-build",
     "dev": "npm run dev:nuxt3",
     "dev:nuxt2": "npm --prefix playground-nuxt2 run dev",
     "dev:nuxt3": "nuxi dev playground-nuxt3",
-    "dev:build": "npm run dev:build:nuxt3",
+    "dev:build": "npm run dev:build:nuxt3 && npm run dev:build:nuxt2",
     "dev:build:nuxt2": "npm --prefix playground-nuxt2 run build",
     "dev:build:nuxt3": "nuxi build playground-nuxt3",
     "dev:prepare": "nuxt-module-build --stub && nuxi prepare playground-nuxt3 && nuxt prepare playground-nuxt2",
