+++ src/runtime/components/vue2/LitElementSsr.vue
@@ -1,0 +1,25 @@
+<template>
+  <component :is="litElementTagName" v-html="litSsrHtml" />
+</template>
+
+<script lang="ts">
+import { renderLitElement } from "../../utils/litRenderer";
+export default {
+  name: "LitElementSsr",
+  data() {
+    const defaultSlot = this.$slots.default?.();
+    const litElementVnode = defaultSlot?.[0];
+    const litElementTagName = litElementVnode?.type;
+
+    return {
+      litElementVnode,
+      litElementTagName,
+      litSsrHtml: ""
+    };
+  },
+
+  async serverPrefetch() {
+    this.litSsrHtml = await renderLitElement(this.litElementTagName, this.litElementVnode, () => {});
+  }
+};
+</script>
